Added Locations List Return for forms
