+++ src/components/animal/AnimalForm.js
@@ -1,33 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import AnimalManager from '../../modules/AnimalManager';
 import { firstLetterCase } from '../../modules/helpers';
-import LocationManager from '../../modules/LocationManager';
+import LocationsReturn from '../location/LocationsReturn';
 import './AnimalForm.css';
-
-const LocationList = () => {
-  // The initial state is an empty array
-  const [locations, setLocations] = useState([]);
-
-  const getLocations = () => {
-    // After the data comes back from the API, we
-    //  use the setLocations function to update state
-    return LocationManager.getAll().then(locationsFromAPI => {
-      setLocations(locationsFromAPI)
-    });
-  };
   
-  console.log(`arrays ${locations} -> ${setLocations}`)
-  // got the locations from the API on the component's first render
-  useEffect(() => {
-    getLocations();
-  }, []);
-
-  // Finally we use map() to "loop over" the locations array to show a list of location cards
-  return (
-      locations
-  );
-}
-
 // Form function for adding animals
 const AnimalForm = props => {
 
@@ -39,7 +15,7 @@ const AnimalForm = props => {
 
   const buildLocations = () => {
     let selectHTML = []
-    LocationList().forEach((location => {
+    LocationsReturn().forEach((location => {
         selectHTML.push(<option value={location.name}>{location.name}</option>)
     }));
     return selectHTML;
@@ -114,4 +90,4 @@ const AnimalForm = props => {
   );
 };
 
-export default AnimalForm
+export default AnimalForm;
