+++ src/components/employee/EmployeeForm.js
@@ -1,37 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { firstLetterCase } from '../../modules/helpers'
 import EmployeeManager from '../../modules/EmployeeManager';
-import LocationManager from '../../modules/LocationManager';
-
+import LocationsReturn from '../location/LocationsReturn';
 
 import './EmployeeForm.css';
-
-
-const LocationList = () => {
-    // The initial state is an empty array
-    const [locations, setLocations] = useState([]);
-  
-    const getLocations = () => {
-      // After the data comes back from the API, we
-      //  use the setLocations function to update state
-      return LocationManager.getAll().then(locationsFromAPI => {
-        setLocations(locationsFromAPI)
-      });
-    };
-  
-    console.log(`arrays ${locations} -> ${setLocations}`)
-    // got the locations from the API on the component's first render
-    useEffect(() => {
-      getLocations();
-    }, []);
-  
-    // Finally we use map() to "loop over" the locations array to show a list of location cards
-    return (
-        locations
-    );
-  }
-
-
 
 const EmployeeForm = props => {
     const [ employee, setEmployee ] = useState({ name: "", experience: "", location: "", picture: "./employee.jpg" });
@@ -39,7 +11,7 @@ const EmployeeForm = props => {
     
     const buildLocations = () => {
         let selectHTML = []
-        LocationList().forEach((location => {
+        LocationsReturn().forEach((location => {
             selectHTML.push(<option value={location.name}>{location.name}</option>)
         }));
         return selectHTML;
