+++ src/components/location/LocationsReturn.js
@@ -1,0 +1,29 @@
+import React, { useState, useEffect } from 'react';
+//import the components we will need
+import LocationCard from './LocationCard';
+import LocationManager from '../../modules/LocationManager';
+
+const LocationsReturn = () => {
+    // The initial state is an empty array
+    const [locations, setLocations] = useState([]);
+
+    const getLocations = () => {
+      // After the data comes back from the API, we
+      //  use the setLocations function to update state
+      return LocationManager.getAll().then(locationsFromAPI => {
+        setLocations(locationsFromAPI)
+      });
+    };
+    
+    console.log(`arrays ${locations} -> ${setLocations}`)
+    // got the locations from the API on the component's first render
+    useEffect(() => {
+      getLocations();
+    }, []);
+  
+    // Finally we use map() to "loop over" the locations array to show a list of location cards
+    return (
+        locations
+    );
+};
+export default LocationsReturn
